fix: Delete article issue
