+++ src/app/(global)/(dashboard)/manage/content/page.tsx
@@ -1,20 +1,44 @@
 'use client';
 
 import { NewspaperIcon } from 'lucide-react';
-import { useQuery } from '@tanstack/react-query';
+import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import TitleWIthIcon from '@/components/dashboard/TitleWIthIcon';
 // import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
 // import { OrderBy } from '@/types';
 // import SearchInput from '@/components/custom/SearchInput';
-import { getUserArticles } from '@/lib/authApi';
+import { deleteArticle, getUserArticles } from '@/lib/authApi';
 import LargeArticleCard from '@/components/article/LargeArticleCard';
 import { QUERY_KEY } from '@/constants';
+import ConfirmDialog from '@/components/custom/ConfirmDialog';
+import { toast } from '@/components/ui/use-toast';
+import { useState } from 'react';
+import { useDialog } from '@/stores/useDialogStore';
 
 export default function ContentPage() {
+  const [deletedArticleId, setDeletedArticleId] = useState('');
+  const { onOpen } = useDialog();
+
   const { data: articles } = useQuery({
     queryKey: [QUERY_KEY.article],
     queryFn: getUserArticles,
   });
+
+  const queryClient = useQueryClient();
+  const { mutate: deleteArticleMutate } = useMutation({
+    mutationFn: deleteArticle,
+    onSuccess: () => {
+      toast({ title: '刪除成功', variant: 'success' });
+      queryClient.invalidateQueries({ queryKey: [QUERY_KEY.article] });
+    },
+    onError: () => {
+      toast({ title: '刪除失敗', description: '請聯繫客服，或稍後再試', variant: 'error' });
+    }
+  });
+
+  function deleteArticleConfirm(articleId: string) {
+    onOpen();
+    setDeletedArticleId(articleId);
+  }
 
   return (
     <>
@@ -35,10 +59,11 @@ export default function ContentPage() {
         <p className='my-5 text-grey-300'>共 {articles?.length} 篇文章</p>
         <div className='flex flex-col gap-4'>
           {articles?.map((article) => (
-            <LargeArticleCard key={article.id} article={article} isUsersArticle />
+            <LargeArticleCard key={article.id} article={article} isUsersArticle deleteArticleConfirm={deleteArticleConfirm} />
           ))}
         </div>
       </section>
+      <ConfirmDialog description='確定要刪除嗎？' onConfirm={() => deleteArticleMutate({ articleId: deletedArticleId })} />
     </>
   );
 }
